requirements.txt
fastapi==0.124.0
Jinja2==3.1.6
python-multipart==0.0.20
SQLAlchemy==2.0.44
uvicorn==0.38.0
python-dateutil==2.8.2
